VSOP improvements and low-accuracy-pendant added
VSOP87 model has been almost completely debugged, hence the computation of Jupiter's as well as the Earth's position in heliocentric-baricentric rectangular coordinates, as well as the conversion to equatorial coordinates from the Earth looks flawless. Nutation correction terms are missing. The VSOP-Model should subsequently implement the E5 -Model for the Galileian satellites.
A low accuracy pendant has been added, it keeps into account the gravitational resonance of the moons and looks like having an acceptable precision.