Added a complte class outside of the pymeeus libarary containg methods to compute all phenomena in the jupiter system. The export of the functions has mainly to do with the fact that they are not strictly part of "Astronomical Algorithms" by Meeus, therefore do not belong in classes related to that Deprecated functions have been removed from the library.
The current phase correction method,the avanzo-centro-regresso works fine for Io and less fine for all other Moons, several attempt at modelling a more accurate behaviour have been made, Nevertheless only improving the anvanzo-centro-regresso model very slightly.They are documented in the new notebook. That is Still a work in progress. It will probably only be solved with a complete model of transients (on which work has begun)